fix: al contraer el AsideBar algunos elementos se movian hacia abajo
